refactor: Router설정 App.tsx로 이동
- Router 설정은 main.tsx보다 최상위 App.tsx에 있는게 자연스러움
- App.tsx가 애플리케이션의 주요 구조와 설정을 담당하는 컴포넌트이기 때움
- 기존의 App.tsx는 MainPage.tsx로 수정함
